% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characteristic_coding.R
\name{characteristic_coding}
\alias{characteristic_coding}
\title{Coding between person characteristics in longitudinal data}
\usage{
characteristic_coding(
  data,
  ItemPrefix,
  RecodeValues = NULL,
  NewName = ItemPrefix,
  Approach,
  PriorityResponse = NULL
)
}
\description{
When data is collected across multiple time points, some characteristics are expected to
be stable over time, but respondents may give inconsistent and contradictory responses.
This function offers an avenue for examining inconsistencies and four methods for
resolving conflicts to produce a single value for each respondent across all time points.
The methods are particularly useful in the presence of check-all-that-apply type responses
at each time point.
}
