% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_codebook.R
\name{create_codebook}
\alias{create_codebook}
\title{Create codebook for a dataset}
\usage{
create_codebook(
  data,
  export.type = NULL,
  export.name = "MyCodebook",
  return.object = TRUE
)
}
\arguments{
\item{data}{a \code{data.frame}}

\item{export.type}{current options are \code{"excel"} and \code{"csv"}}

\item{export.name}{character string indicating file path and file name for exported codebook. File extention is added by the function and should not be included here.}

\item{return.object}{should the codebook dataframe be returned?}
}
\value{
data.frame
}
\description{
The function extracts and organizes variable attributes into a dataframe that can be used for other R functions or exported.
}
\details{
If \code{export.type = NULL} & \code{return.object = FALSE}, the function does not output anything.
}
