% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_codebook.R
\name{get_labels}
\alias{get_labels}
\title{Extracting labels from factor and labelled variables}
\usage{
get_labels(data)
}
\arguments{
\item{data}{a \code{data.frame} with observations in rows and variables in columns}
}
\value{
a \code{data.frame} with a row for each value-label pair
}
\description{
Extracting labels from factor and labelled variables
}
