% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplus_sb2001.R
\name{mplus_sb2001}
\alias{mplus_sb2001}
\title{Satorra-Bentler (2001) Scaled Chisquare}
\usage{
mplus_sb2001(m0, m1, add.index = TRUE, digits = 2)
}
\arguments{
\item{m0}{nested model}

\item{m1}{comparison model}

\item{add.index}{should change (delta) in fit index values (CFI, RMSEA, SRMR) be added to the output?}
}
\value{
single row dataframe
}
\description{
Calculate the Satorra-Bentler (2001) scaled (mean-adjusted) chisquare difference test
for nested models run in Mplus and imported using MPlusAutomation.
Models should be estimated with MLM, MLR, or WLSM (https://statmodel.com/chidiff.shtml)
}
