% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_score.R
\name{scale_score}
\alias{scale_score}
\title{Calculate scale score}
\usage{
scale_score(data, items = names(data), type = c("sum", "mean"), min.valid = 1)
}
\arguments{
\item{data}{a \code{data.frame}}

\item{items}{character vector of column names in \code{data} to use in the scale}

\item{type}{should the score be the \code{"sum"} or the \code{"mean"} of item responses?}

\item{min.valid}{the minimum number of valid responses to receive a score. The default is 1, otherwise NA is returned.}
}
\value{
numeric vector
}
\description{
Combine item responses into a scale score
}
\examples{
set.seed(9876)
data <- data.frame(x1 = sample(0:4, 20, replace = TRUE),
                 x2 = sample(0:4, 20, replace = TRUE),
                 x3 = sample(0:4, 20, replace = TRUE),
                 x4 = sample(0:4, 20, replace = TRUE))

sum.score <- scale_score(data, paste0("x", 1:4), type = "sum")

## Incorporate missing data
dfmiss <- apply(data, 2, function(x){
                x[sample(c(1:nrow(df)), floor(nrow(df)/10))] <- NA
                x})
dfmiss <- as.data.frame(dfmiss)

# Score NA if missing on more than 1 item
mean.score <- scale_score(dfmiss, paste0("x", 1:4), type = "mean", min.valid = 3)

}
